Always close the abort eviction channel
Previously we closed the abort channel if we hit our eviction deadline, but not
when we encountered an unexpected error evicting a pod. This means we'd leak a
goroutine if eviction A returned an unexpected error while eviction B was
spinning forever trying to evict a pod (due for example to an unsatisfiable pod
disruption budget).