Make pod eviction filters configurable
This changes the default behaviour in that Draino would previously evict
unreplicated pods and pods with local storage, and now will not. This is
not ideal, but https://github.com/alecthomas/kingpin/issues/243 would
result in a confusing user experience if we used a mix of default true and
default false boolean flags. Given the project is about 10 minutes old it
seems acceptable to make this change.