Fix data race in drain scheduler
This commit fixes a data race when calling HasSchedule. This happens because
the `time.AfterFunc` function is called by a new goroutine and it updates
fields which may be read in parallel by `HasSchedule`.

The `running` return value from `HasSchedule` was removed to simplify the code
as it wasn't used anywhere.