Remplace "," with "." to be compliant with Label regexp
In case the node is eligible for 2 config the label value was using a "," which is not accepted.
```
"error":"Node \"ip-10-49-244-120.ec2.internal\" is invalid: metadata.labels: Invalid value: \"database-with-local-data,standard\": a valid label must be an empty string or consist of alphanumeric characters, '-', '_' or '.', and must start and end with an alphanumeric character (e.g. 'MyValue',  or 'my_value',  or '12345', regex used for validation is '(([A-Za-z0-9][-A-Za-z0-9_.]*)?[A-Za-z0-9])?')"
```