feature: sync is pod ready function with k8s (#252)
We were seeing different behaviours of direct k8s evictions vs. our simulation tests. The origin is that both have different ideas of if a pod is ready or not.
To have a consistent view, this PR will use the kubernetes package util to validate if a pod is ready or not.